Try to prevent pygments from reading files
Since the pygmentize script and the pygments server probably run on
different servers, the ability to read local files isn’t very useful
here; on the other hand, it may reveal the contents of files that are
readable to pygments-server but shouldn’t be public (as happened in
T286724 [1]). We try to protect against this by overriding Python’s
builtin open() function with one that always returns an empty IO object;
however, this is not foolproof (e.g. if pygments switches to os.open()
for some reason), so also add a note to the README about how people
should configure their web server.

[1]: https://phabricator.wikimedia.org/T286724